added timing system

+++ BinSearch.java
@@ -46,7 +46,7 @@ public class BinSearch
       m = (lo + hi) / 2;
 
       // target found
-      if (a[m] == target) {
+      if ((int) a[m] == (int) target) {
         return m;
       }
